fix(config): unify config_setting known-key set + pay down #3074 cold-reader debt (#3127)
Issue A: `config_setting get`/`set` validated a key against the
OVERLAY_OVERRIDABLE / REGISTRY / COLD registries only, excluding the
COLD_HOOK_SETTINGS gate keys (e.g. out_of_band_merge_gate_enabled) that
`list` displays — so get/list/set/clear disagreed on "known key". Add
COLD_HOOK_SETTINGS to the single allowlist so get accepts (and set/clear
handle) every key list can show; get now reports a gate key's in-code
ColdHookSetting default when no DB row exists, instead of refusing.

Issue B (#3074 deferred debt):
- Rebalance cold_reader.py (>10 module-level fns): split the Django-free
  raw sqlite plumbing (path resolution, read-only open, single-row fetch,
  loop_status, row_exists) into a new teatree.config.cold_db module,
  re-exported through cold_reader so every cold_reader.<name> reference is
  untouched. Behaviour preserved exactly; no Django import at module scope.
  cold_reader 8 public fns, cold_db 4.
- Justify the bare `# noqa: PLC0415` deferred cold_reader imports in
  config_migration / timeouts / availability (deferred to keep the module
  import light — importing the teatree.config facade eagerly loads its heavy
  package __init__).
- dict[str, object] casts: audited the ~11 #3074 left; none converted to
  TypedDict — see PR description.

Open questions & assumptions:
- assumed: get on a cold-hook gate key with no DB row reports the in-code
  ColdHookSetting default (source "code default"); this matches what the
  cold reader actually resolves.
- assumed: the shared single-row fetch was promoted _fetch_one -> public
  cold_db.fetch_one since it is now a cross-module helper.
- decided: 0 dict[str,object]->TypedDict conversions — the #3074 casts are
  all heterogeneous JSON/overlay-registry/host-API maps where dict[str,object]
  is the honest type, and the only fixed-shape candidate (pr.check_gates)
  sits in an over-module-health-cap file where adding a TypedDict would trip
  the shrink-ratchet.

Closes #3127
